+++ dapp-blackjack/src/components/cards/Card5.tsx
@@ -1,0 +1,15 @@
+import { CardProp } from "./CardProp";
+import { Suit, getColorClassName } from "./Suit";
+
+export function Card5({ suit }: CardProp) {
+    return (<div className="card">
+        <div className={`front ${getColorClassName(suit)}`}>
+            <div className="index">5<br /><Suit name={suit} /></div>
+            <div className="spotA1"><Suit name={suit} /></div>
+            <div className="spotA5"><Suit name={suit} /></div>
+            <div className="spotB3"><Suit name={suit} /></div>
+            <div className="spotC1"><Suit name={suit} /></div>
+            <div className="spotC5"><Suit name={suit} /></div>
+        </div>
+    </div>)
+}