+++ dapp-blackjack/src/components/cards/CardProp.ts
@@ -1,0 +1,5 @@
+import { SuitType } from "./Suit"
+
+export type CardProp = {
+    suit: SuitType
+}