+++ dapp-contract-blackjack/src/rollups.rs
@@ -3,9 +3,11 @@ pub mod rollup {
         body::to_bytes, client::HttpConnector, header, Body, Client, Method, Request, Response,
         StatusCode,
     };
+    use hyper_tls::HttpsConnector;
     use serde_json::{from_str, json, Value};
     use std::{env, error::Error, str::from_utf8, sync::Arc, time::Duration};
     use tokio::sync::{mpsc::Sender, Mutex};
+    use std::fs;
 
     use crate::{
         models::{
@@ -23,7 +25,9 @@ pub mod rollup {
     ) -> Result<(), Box<dyn Error>> {
         println!("Starting loop...");
 
-        let client = Client::new();
+        // let client = Client::new();
+        let https = HttpsConnector::new();
+        let client = Client::builder().build::<_, hyper::Body>(https);
         let server_addr = env::var("MIDDLEWARE_HTTP_SERVER_URL")?;
 
         let mut status = "accept";
@@ -52,12 +56,10 @@ pub mod rollup {
 
                 status = match request_type {
                     "advance_state" => {
-                        handle_advance(manager.clone(), &client, &server_addr[..], body, sender)
-                            .await?
+                        handle_advance(manager.clone(), &server_addr[..], body, sender).await?
                     }
                     "inspect_state" => {
-                        handle_inspect(manager.clone(), &client, &server_addr[..], body, sender)
-                            .await?
+                        handle_inspect(manager.clone(), &server_addr[..], body, sender).await?
                     }
                     &_ => {
                         eprintln!("Unknown request type");
@@ -65,13 +67,13 @@ pub mod rollup {
                     }
                 }
             }
+            println!("waiting 5s...");
             tokio::time::sleep(Duration::from_secs(5)).await;
         }
     }
 
     async fn handle_inspect(
         manager: Arc<Mutex<Manager>>,
-        client: &Client<HttpConnector>,
         server_addr: &str,
         body: Value,
         sender: &Sender<Value>,
@@ -86,6 +88,25 @@ pub mod rollup {
         let payload = get_payload_from_root(&body).ok_or("Invalid payload")?;
         let action = get_from_payload_action(&payload);
         match action.as_deref() {
+            Some("show_player") => {
+                let input = payload.get("input").ok_or("Invalid field input")?;
+
+                // Parsing JSON
+                let address = input
+                    .get("address")
+                    .ok_or("Invalid field address")?
+                    .as_str()
+                    .ok_or("Invalid address")?;
+                let address_owner = address.trim_start_matches("0x");
+                let address_encoded = bs58::encode(address_owner).into_string();
+
+                let address_path = format!("./data/address/{}.json", address_encoded);
+                let content = fs::read_to_string(address_path).expect("Open player file error");
+                let json_as_hex = hex::encode(content.to_string());
+                let report = json!({ "payload": format!("0x{}", json_as_hex) });
+                println!("Report: {:}", report);
+                let _ = send_report(report.clone()).await;
+            }
             Some("show_games") => {
                 let manager = manager.lock().await;
                 let games = manager.show_games_id_available();
@@ -143,7 +164,7 @@ pub mod rollup {
                 let scoreboard = manager
                     .get_scoreboard(table_id, game_id)
                     .ok_or("Scoreboard not found searching by table_id")?;
-                
+
                 let response = json!({
                     "scoreboard": scoreboard.to_json(),
                 });
@@ -161,7 +182,6 @@ pub mod rollup {
 
     async fn handle_advance(
         manager: Arc<Mutex<Manager>>,
-        client: &Client<HttpConnector>,
         server_addr: &str,
         body: Value,
         sender: &Sender<Value>,
@@ -169,10 +189,10 @@ pub mod rollup {
         println!("Handling advance");
 
         println!("body {:}", &body);
-        let run_async = false;
-        if run_async {
+        let run_async = std::env::var("RUN_GAME_ASYNC").unwrap_or("true".to_string());
+        if run_async == "true" {
             sender.send(body).await?;
-            return Ok("accept")
+            return Ok("accept");
         }
         let payload = get_payload_from_root(&body).ok_or("Invalid payload")?;
         let action = get_from_payload_action(&payload);
@@ -340,7 +360,9 @@ pub mod rollup {
         report: Value,
     ) -> Result<&'static str, Box<dyn std::error::Error>> {
         let server_addr = std::env::var("ROLLUP_HTTP_SERVER_URL")?;
-        let client = hyper::Client::new();
+        // let client = hyper::Client::new();
+        let https = HttpsConnector::new();
+        let client = Client::builder().build::<_, hyper::Body>(https);
         let req = hyper::Request::builder()
             .method(hyper::Method::POST)
             .header(hyper::header::CONTENT_TYPE, "application/json")